Feat: Add role and ownership based security to API

+++ src/main/java/com/example/mefitapp/controllers/ExerciseController.java
@@ -4,10 +4,11 @@ import com.example.mefitapp.models.Exercise;
 import com.example.mefitapp.models.ExerciseSet;
 import com.example.mefitapp.repositories.ExerciseRepository;
 import com.example.mefitapp.services.ExerciseService;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.example.mefitapp.services.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.util.ReflectionUtils;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,7 +29,7 @@ public class ExerciseController {
     private ExerciseService exerciseService;
 
     @Autowired
-    private ObjectMapper objectMapper;
+    private SecurityService securityService;
 
     @GetMapping()
     public ResponseEntity<List<Exercise>> getAllExercises() {
@@ -65,42 +66,57 @@ public class ExerciseController {
     }
 
     @PostMapping
-    public ResponseEntity<Exercise> addExercise(@RequestBody Exercise exercise) {
-        Exercise returnExercise = exerciseRepository.save(exercise);
-        HttpStatus status = HttpStatus.CREATED;
-        return new ResponseEntity<>(returnExercise, status);
+    public ResponseEntity<Exercise> addExercise(@RequestBody Exercise exercise, Authentication authentication) {
+        if (securityService.isContributor(authentication)) {
+            Exercise returnExercise = exerciseRepository.save(exercise);
+            HttpStatus status = HttpStatus.CREATED;
+            return new ResponseEntity<>(returnExercise, status);
+        } else {
+            HttpStatus status = HttpStatus.UNAUTHORIZED;
+            return new ResponseEntity<>(status);
+        }
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<Exercise> updateExercise(@PathVariable Long id, @RequestBody Map<String, Object> updates) {
+    public ResponseEntity<Exercise> updateExercise(@PathVariable Long id,
+           @RequestBody Map<String, Object> updates, Authentication authentication) {
         Exercise returnExercise = new Exercise();
         HttpStatus status;
-        if (exerciseRepository.existsById(id)) {
-            status = HttpStatus.OK;
-            Exercise toBePatchedExercise = exerciseRepository.findById(id).get();
-            // Map key is field name, v is value
-            updates.forEach((k, v) -> {
-                // use reflection to get field k on exercise and set it to value v
-                Field field = ReflectionUtils.findField(Exercise.class, k);
-                field.setAccessible(true);
-                ReflectionUtils.setField(field, toBePatchedExercise, v);
-            });
-            exerciseRepository.save(toBePatchedExercise);
-            returnExercise = toBePatchedExercise;
+        if (securityService.isContributor(authentication)) {
+            if (exerciseRepository.existsById(id)) {
+                status = HttpStatus.OK;
+                Exercise toBePatchedExercise = exerciseRepository.findById(id).get();
+                // Map key is field name, v is value
+                updates.forEach((k, v) -> {
+                    // use reflection to get field k on exercise and set it to value v
+                    Field field = ReflectionUtils.findField(Exercise.class, k);
+                    field.setAccessible(true);
+                    ReflectionUtils.setField(field, toBePatchedExercise, v);
+                });
+                exerciseRepository.save(toBePatchedExercise);
+                returnExercise = toBePatchedExercise;
+            } else {
+                status = HttpStatus.NOT_FOUND;
+            }
+            return new ResponseEntity<>(returnExercise, status);
         } else {
-            status = HttpStatus.NOT_FOUND;
+            status = HttpStatus.UNAUTHORIZED;
+            return new ResponseEntity<>(status);
         }
-        return new ResponseEntity<>(returnExercise, status);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteExercise(@PathVariable Long id) {
+    public ResponseEntity<HttpStatus> deleteExercise(@PathVariable Long id, Authentication authentication) {
         HttpStatus status;
-        if (exerciseRepository.existsById(id)) {
-            status = HttpStatus.OK;
-            exerciseRepository.deleteById(id);
+        if (securityService.isAdmin(authentication)) {
+            if (exerciseRepository.existsById(id)) {
+                status = HttpStatus.OK;
+                exerciseRepository.deleteById(id);
+            } else {
+                status = HttpStatus.NOT_FOUND;
+            }
         } else {
-            status = HttpStatus.NOT_FOUND;
+            status = HttpStatus.UNAUTHORIZED;
         }
         return new ResponseEntity<>(status);
     }
